Moving GetMachineOnlineStatusAsync from core to client.
